core/standards/scripts/scheduler2/location_hash/401.html

<!doctype html>
<title>Location hash - set hash after page completely loaded</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test();
var histlength;
var step_delay = 100;
t.step(function() {
    if( location.hash && location.hash != '#' ) {
        location.href = location.href.replace(/#.*$/,'');
        return;
    }
    histlength = history.length;
});

window.onload = function () {
    setTimeout(step1, step_delay);  
};

function step1() {
    t.step(function() {
        location.hash = '#foo';
    });
    setTimeout(last_step, step_delay);
}

function last_step() {
    t.step(function() {
        assert_equals( location.hash, '#foo' );
        assert_equals( history.length, histlength + 1 );
    });
    t.done();
}
</script>
<div id="log"></div>
